requirements.txt
yt_dlp
ffmpeg-python
music-tag
